+++ forui/lib/foundation.dart
@@ -1,10 +1,11 @@
 /// Low-level utilities and services.
 library forui.foundation;
 
+export 'src/foundation/accessibility.dart';
 export 'src/foundation/barrier.dart';
 export 'src/foundation/collapsible.dart';
 export 'src/foundation/doc_templates.dart' hide Control, Focus, FormFieldKey, Scroll, Semantics, TappableGroup;
-export 'src/foundation/focused_outline.dart' hide FocusHighlightScope;
+export 'src/foundation/focused_outline.dart';
 export 'src/foundation/form_field_properties.dart';
 export 'src/foundation/notifiers.dart' hide InternalFMultiValueControl;
 export 'src/foundation/rendering.dart' hide Alignments, DefaultData, RenderBoxes;